+++ packages/editor/src/lib/primitives/geometry/Arc2d.ts
@@ -89,7 +89,7 @@ export class Arc2d extends Geometry2d {
 		return vertices
 	}
 
-	toSvg(first = true) {
+	getSvgPathData(first = true) {
 		const { start, end, radius, largeArcFlag, sweepFlag } = this
 		return `${first ? `M${start.toFixed()}` : ``} A${radius} ${radius} 0 ${largeArcFlag} ${sweepFlag} ${end.toFixed()}`
 	}
