+++ packages/editor/src/lib/primitives/geometry/Group2d.ts
@@ -96,4 +96,8 @@ export class Group2d extends Geometry2d {
 		}
 		return path
 	}
+
+	getSvgPathData(): string {
+		return this.children.map((c, i) => (c.isLabel ? '' : c.getSvgPathData(i === 0))).join(' ')
+	}
 }
