--- packages/tldraw/src/lib/shapes/geo/components/SolidStyleHeart.tsx
@@ -1,23 +1,0 @@
-import { TLGeoShape } from '@tldraw/editor'
-import * as React from 'react'
-import { ShapeFill, useDefaultColorTheme } from '../../shared/ShapeFill'
-import { getHeartPath } from '../geo-shape-helpers'
-
-export const SolidStyleHeart = React.memo(function SolidStyleHeart({
-	w,
-	h,
-	strokeWidth: sw,
-	fill,
-	color,
-}: Pick<TLGeoShape['props'], 'w' | 'h' | 'fill' | 'color'> & {
-	strokeWidth: number
-}) {
-	const theme = useDefaultColorTheme()
-	const d = getHeartPath(w, h)
-	return (
-		<>
-			<ShapeFill d={d} color={color} fill={fill} theme={theme} />
-			<path d={d} stroke={theme[color].solid} strokeWidth={sw} fill="none" />
-		</>
-	)
-})
